\name{mmSAR}
\Rdversion{0.1}
\alias{mmSAR}
\alias{mmSAR}
\docType{package}
\title{
~~ mmSAR is an R package for the modelling of the Species-Area Relationship (SAR). ~~
}
\description{
~~ The mmSAR package implement the multimodel SAR paradigm and provides ecologists with user friendly functions fit SAR models and calculate multimodel SAR inference. ~~
}
\details{
\tabular{ll}{
Package: \tab mmSAR\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2009-09-11\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
~~ François Guilhaumon <francoisguilhaumon@gmail.com> ~~
}
\references{
~~ Burnham, K. P., and D. R. Anderson.  2002.  Model selection and multimodel inference: a practical information-theoretic approach. 2nd Edition. Springer-Verlag, New York, New York, USA.  488 pp. ~~
}

\keyword{ package }
\seealso{
}
\examples{

}
